# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Erro de domínio das rotas `/auth`, `/me` e `/admin` (ADR-008).",
        "example": {
          "error": "user_not_found",
          "message": "Usuário não encontrado."
        },
        "properties": {
          "error": {
            "description": "Slug estável do erro, ex.: `user_not_found`.",
            "title": "Error",
            "type": "string"
          },
          "message": {
            "description": "Mensagem legível, em português.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_AnalysisSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[AnalysisSummary]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerToken": {
        "description": "Token opaco emitido por `POST /auth/token`. Cole apenas o token (o Swagger acrescenta o prefixo `Bearer`).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API HTTP do servidor de análise de dados via MCP: emissão de tokens (`/auth`), dados do próprio\nusuário (`/me`), administração (`/admin/*`) e `/health`.\n\n**Autenticação.** Emita um token em `POST /auth/token` (e-mail + senha), clique em **Authorize**\ne cole só o token. `/me` exige um token válido; `/admin/*` exige também administrador.\n\n**Erros.** As rotas de `/auth`, `/me` e `/admin` respondem erros de domínio no formato\n`{\"error\": \"<slug>\", \"message\": \"...\"}`. Erros de validação do corpo ou dos parâmetros vêm no\nformato padrão do FastAPI (`422`, `{\"detail\": [...]}`).\n\n**O `/mcp` não está neste documento.** O endpoint MCP (Streamable HTTP/JSON-RPC) é descrito em\n`docs/MCP.md`.\n",
    "title": "Análise de Dados Genérica com MCP — API HTTP",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/analyses": {
      "get": {
        "description": "Lista paginada, com busca (`q`, nome ou descrição) e filtros por data source e ativo.",
        "operationId": "list_analyses_admin_analyses_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "data_source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Source Id"
            }
          },
          {
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AnalysisSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Listar analyses",
        "tags": [
          "admin-analyses"
        ]
      },
      "post": {
        "description": "Cria a análise, a query (step) e os perfis numa transação, após validação estática da definição (parâmetros, SQL somente SELECT, `cache_frequency`).",
        "operationId": "create_analysis_admin_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito. Slugs: `analysis_name_already_exists`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Create Analysis Admin Analyses Post"
                }
              }
            },
            "description": "Entidade não processável. Slugs: `invalid_reference`, `invalid_analysis_definition`. Também pode vir no formato padrão do FastAPI (`{\"detail\": [...]}`) quando o corpo ou os parâmetros não passam na validação."
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Criar análise",
        "tags": [
          "admin-analyses"
        ]
      }
    },
    "/admin/analyses/{analysis_id}": {
      "delete": {
        "description": "Exclusão física, só sem histórico de execuções; com histórico responde 409 e o caminho é desativar (`is_active: false`).",
        "operationId": "delete_analysis_admin_analyses__analysis_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `analysis_not_found`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito. Slugs: `analysis_has_history`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Excluir análise",
        "tags": [
          "admin-analyses"
        ]
      },
      "get": {
        "description": "Devolve a análise com parâmetros, query e perfis.",
        "operationId": "get_analysis_admin_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `analysis_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Detalhar análise",
        "tags": [
          "admin-analyses"
        ]
      },
      "patch": {
        "description": "Altera apenas os campos enviados. `parameters` substitui o objeto inteiro; `step` mescla `sql`/`params`. A definição só é revalidada quando o PATCH a toca; alterações de definição atualizam `updated_at` (invalida o cache).",
        "operationId": "update_analysis_admin_analyses__analysis_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `analysis_not_found`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito. Slugs: `analysis_name_already_exists`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Update Analysis Admin Analyses  Analysis Id  Patch"
                }
              }
            },
            "description": "Entidade não processável. Slugs: `invalid_reference`, `invalid_analysis_definition`. Também pode vir no formato padrão do FastAPI (`{\"detail\": [...]}`) quando o corpo ou os parâmetros não passam na validação."
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Alterar análise",
        "tags": [
          "admin-analyses"
        ]
      }
    },
    "/admin/analyses/{analysis_id}/cache/invalidate": {
      "post": {
        "description": "Invalida o cache da análise atualizando `updated_at = NOW()`.",
        "operationId": "invalidate_analysis_cache_admin_analyses__analysis_id__cache_invalidate_post",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `analysis_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Invalidar cache",
        "tags": [
          "admin-analyses"
        ]
      }
    },
    "/admin/analyses/{analysis_id}/executions": {
      "get": {
        "description": "Atalho de `GET /admin/executions` já filtrado pela análise (aceita também as análises inativas). Mesmos filtros, sem `analysis_id`.",
        "operationId": "list_analysis_executions_admin_analyses__analysis_id__executions_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "success",
                    "volume_exceeded",
                    "error",
                    "timeout"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "error_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ANALYSIS_NOT_FOUND",
                    "INVALID_PARAMETERS",
                    "INVALID_ANALYSIS_CONFIG",
                    "DATA_SOURCE_UNAVAILABLE",
                    "QUERY_TIMEOUT",
                    "QUERY_FAILED",
                    "INTERNAL_ERROR"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Code"
            }
          },
          {
            "in": "query",
            "name": "cached",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cached"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "min_time_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Time Ms"
            }
          },
          {
            "in": "query",
            "name": "parameters",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parameters"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExecutionSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `analysis_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 List Analysis Executions Admin Analyses  Analysis Id  Executions Get"
                }
              }
            },
            "description": "Entidade não processável. Slugs: `invalid_period`, `invalid_parameters_filter`. Também pode vir no formato padrão do FastAPI (`{\"detail\": [...]}`) quando o corpo ou os parâmetros não passam na validação."
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Histórico da análise",
        "tags": [
          "admin-analyses"
        ]
      }
    },
    "/admin/analyses/{analysis_id}/profiles": {
      "put": {
        "description": "Substitui o conjunto de perfis que podem executar a análise.",
        "operationId": "set_analysis_profiles_admin_analyses__analysis_id__profiles_put",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIdsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `analysis_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Set Analysis Profiles Admin Analyses  Analysis Id  Profiles Put"
                }
              }
            },
            "description": "Entidade não processável. Slugs: `invalid_reference`. Também pode vir no formato padrão do FastAPI (`{\"detail\": [...]}`) quando o corpo ou os parâmetros não passam na validação."
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Definir perfis da análise",
        "tags": [
          "admin-analyses"
        ]
      }
    },
...
  },
...
}